server/requirements.txt
Flask==1.0.2
Flask-SQLAlchemy
marshmallow
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
google-cloud-datastore==1.9.0
pyjwt==1.7.1
requests==2.22.0
